requirements.txt
Flask==3.1.2
flask-cors==6.0.1
google-generativeai==0.8.5
numpy==2.3.5
